@@ -4,14 +4,14 @@ version: 1.0.14
 publish_to: none
 
 environment:
-  sdk: ">=2.15.0 <3.0.0"
+  sdk: ">=2.19.0 <3.0.0"
 
 dependencies:
   auto_route: ^5.0.4
   auto_size_text: ^3.0.0
   collection: ^1.17.0
   dartz: ^0.10.1
-  diox: ^5.0.0
+  dio: ^5.0.0
   flutter_svg: ^1.1.6
   websafe_svg: ^2.1.3+1
   flex_color_scheme: ^6.1.2
@@ -52,6 +52,7 @@ dependencies:
 
 
 dev_dependencies:
+  flutter_phoenix: ^1.1.0
   import_sorter: ^4.6.0
   auto_route_generator: ^5.0.3
   alchemist: ^0.6.0
@@ -66,10 +67,16 @@ dev_dependencies:
   mocktail: ^0.3.0
   very_good_analysis: ^4.0.0+1
   test: ^1.19.5
+  patrol: ^1.0.0
 
 
 scripts: derry.yaml
-
+patrol:
+  app_name: Joyful Noise
+  android:
+    package_name: com.jparrack.joyful_noise.RunnerUITests
+  ios:
+    bundle_id: com.jparrack.joyful-noise.RunnerUITests
 
 flutter:
   uses-material-design: true
@@ -78,3 +85,4 @@ flutter:
     - .env.staging
     - .env.dev
     - .env.production
+    - .env
