@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:shimmer/shimmer.dart';
 
 class LoadingSongTile extends StatelessWidget {
-  const LoadingSongTile({Key? key}) : super(key: key);
+  const LoadingSongTile({super.key});
 
   @override
   Widget build(BuildContext context) {
