@@ -1,5 +1,4 @@
 // Dart imports:
-import 'dart:io';
 
 // Flutter imports:
 import 'package:flutter/material.dart';
@@ -17,7 +16,7 @@ import 'package:joyful_noise/core/presentation/routes/app_router.gr.dart';
 import 'package:joyful_noise/search/presentation/search_bar.dart';
 
 class FavoriteSongsPage extends ConsumerStatefulWidget {
-  const FavoriteSongsPage({Key? key}) : super(key: key);
+  const FavoriteSongsPage({super.key});
 
   @override
   FavoriteSongsPageState createState() => FavoriteSongsPageState();
@@ -26,12 +25,6 @@ class FavoriteSongsPage extends ConsumerStatefulWidget {
 class FavoriteSongsPageState extends ConsumerState<FavoriteSongsPage> {
   @override
   void initState() {
-    if (!Platform.environment.containsKey('FLUTTER_TEST')) {
-      // Just for logging dont want to test this.
-      // coverage:ignore-start
-      ref.read(userNotifierProvider.notifier).getUserPage();
-      // coverage:ignore-end
-    }
     ref.read(favoriteSongsNotifierProvider.notifier).getNextFavoriteSongsPage();
     super.initState();
   }
