@@ -4,7 +4,7 @@ import 'package:flutter/material.dart';
 // Package imports:
 import 'package:auto_route/auto_route.dart';
 import 'package:dartz/dartz.dart';
-import 'package:diox/diox.dart';
+import 'package:dio/dio.dart';
 import 'package:flex_color_scheme/flex_color_scheme.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:newrelic_mobile/newrelic_navigation_observer.dart';
@@ -35,7 +35,7 @@ final initializationProvider = FutureProvider<Unit>(
 );
 
 class AppWidget extends ConsumerWidget {
-  AppWidget({Key? key}) : super(key: key);
+  AppWidget({super.key});
   final _appRouter = AppRouter();
   @override
   Widget build(BuildContext context, WidgetRef ref) {
