@@ -8,8 +8,8 @@ class NoResultsDisplay extends StatelessWidget {
   final String message;
   const NoResultsDisplay({
     required this.message,
-    Key? key,
-  }) : super(key: key);
+    super.key,
+  });
 
   @override
   Widget build(BuildContext context) {
