@@ -32,7 +32,7 @@ class MockPlaylistSongRepository extends Mock implements PlaylistSongsRepository
 class MockUserRepository extends Mock implements UserRepository {}
 
 class FakeUserNotifier extends UserNotifier {
-  FakeUserNotifier(UserRepository userRepository) : super(userRepository);
+  FakeUserNotifier(super.userRepository);
 
   @override
   Future<void> getUserPage() async {
