@@ -39,7 +39,7 @@ class MockUserRepository extends Mock implements UserRepository {}
 class MockNavigatorObserver extends Mock implements NavigatorObserver {}
 
 class FakeUserNotifier extends UserNotifier {
-  FakeUserNotifier(UserRepository userRepository) : super(userRepository);
+  FakeUserNotifier(super.userRepository);
 
   @override
   Future<void> getUserPage() async {
