@@ -35,7 +35,7 @@ class MockFavoriteSongRepository extends Mock implements FavoriteSongsRepository
 class MockUserRepository extends Mock implements UserRepository {}
 
 class FakeUserNotifier extends UserNotifier {
-  FakeUserNotifier(UserRepository userRepository) : super(userRepository);
+  FakeUserNotifier(super.userRepository);
 
   @override
   Future<void> getUserPage() async {
